+++ src/net/sourceforge/kolmafia/request/UseSkillRequest.java
@@ -1323,7 +1323,7 @@ public class UseSkillRequest extends GenericRequest implements Comparable<UseSki
       ++rv;
     }
 
-    rv += KoLCharacter.currentNumericModifier(Modifiers.ADDITIONAL_SONG);
+    rv += (int) KoLCharacter.currentNumericModifier(Modifiers.ADDITIONAL_SONG);
 
     return rv;
   }
@@ -2213,7 +2213,7 @@ public class UseSkillRequest extends GenericRequest implements Comparable<UseSki
         // at a time.
         count = 1;
       }
-      mpCost = 0;
+      mpCost = 0L;
     }
 
     switch (skillId) {
